src/pipelines/components/scraper_components/pd_embed_text/requirements.txt
fondant==0.11.2
fondant[component]
google-cloud-aiplatform==1.34.0
langchain==0.0.329
openai==0.28.1
retry==0.9.2
sentence-transformers==2.2.2
tiktoken==0.5.1